requirements.txt
Flask>=2.0
openai>=1.0
PyPDF2>=3.0
python-dotenv>=0.15
